+++ tests/testthat/test-ronda148.R
@@ -32,6 +32,8 @@ library(DBI)
   x$resumen_tabla$meta$instrumentacion <- NULL
   x$resumen_tabla$meta$derrame <- NULL
   x$resumen_tabla$meta$costo_distintos <- NULL
+  x$resumen_tabla$meta$costo_moda <- NULL
+  x$resumen_tabla$meta$costo_mediana <- NULL
   x
 }