+++ tests/testthat/test-ronda152.R
@@ -41,6 +41,8 @@ skip_if_not_installed("RSQLite")
   x$resumen_tabla$meta$instrumentacion <- NULL
   x$resumen_tabla$meta$derrame <- NULL
   x$resumen_tabla$meta$costo_distintos <- NULL
+  x$resumen_tabla$meta$costo_moda <- NULL
+  x$resumen_tabla$meta$costo_mediana <- NULL
   x
 }