+++ tests/testthat/test-ronda153.R
@@ -67,6 +67,8 @@ setMethod(
   x$resumen_tabla$meta$instrumentacion <- NULL
   x$resumen_tabla$meta$derrame <- NULL
   x$resumen_tabla$meta$costo_distintos <- NULL
+  x$resumen_tabla$meta$costo_moda <- NULL
+  x$resumen_tabla$meta$costo_mediana <- NULL
   x
 }